requirements.txt
anyio==4.7.0
certifi==2024.12.14
colorama==0.4.6
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
python-telegram-bot==21.9
sniffio==1.3.1
tomli==2.2.1
typing-extensions==4.12.2
